*: Modified to conform to Module aware 'go list'
Go 1.13 is module aware, so the output of 'go list' results in paths pointing to module cache (GOPATH/pkg/mod).
Non module aware go versions, the `Root` field in the `go list` output contains `GOPATH`. In module aware Go versions, it points to the root of module in module cache.
Earlier versions, the LICENSEs could be read by concatenating the `GOPATH` and `ImportPath` fields. Now the LICENSES need to be read by using `Root` and `Dir` fields which point to the root of the module and imported folder within the module.

This change fixes bill-of-materials test fails in etcd CI with Go 1.13.
https://github.com/etcd-io/etcd/issues/11132